move src/components/Settings to src/app
